requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
certifi @ file:///C:/b/abs_8a944p1_gn/croot/certifi_1738623753421/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.0
cycler==0.12.1
dash==3.0.4
decorator==5.2.1
# Editable install with no version control (depth_pro==0.1)
-e d:\senior design\structure-from-motion-\3d_construction\ml-depth-pro
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.5.1
gdown==5.2.0
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
huggingface-hub==0.32.4
humanfriendly==10.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio==2.37.0
imgviz==1.7.6
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.18.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
labelme==5.8.1
lazy_loader==0.4
loguru==0.7.3
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.9.4
matplotlib-inline==0.1.7
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
narwhals==1.41.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy==1.26.4
onnxruntime==1.19.2
open3d==0.19.0
opencv-python==4.11.0.86
osam==0.2.3
packaging==24.2
parso==0.8.4
pillow @ file:///C:/b/abs_68t826txdy/croot/pillow_1744613085333/work
pillow_heif==0.22.0
platformdirs==4.3.8
plotly==6.1.2
prompt_toolkit==3.0.51
protobuf==6.30.2
pure_eval==0.2.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==310
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
referencing==0.36.2
regex==2024.11.6
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
retrying==1.3.4
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
tifffile==2024.8.30
timm==1.0.15
tokenizers==0.21.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typing-inspection==0.4.0
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
win32_setctime==1.2.0
zipp==3.21.0
